split into callLambdaSync(), callLambdaAsync(), callLambdaStreaming()
