+++ packages/lambda/src/api/render-media-on-lambda.ts
@@ -19,7 +19,7 @@ import type {
 import {ServerlessRoutines} from '@remotion/serverless/client';
 import type {AwsProvider} from '../functions/aws-implementation';
 import type {AwsRegion} from '../regions';
-import {callLambda} from '../shared/call-lambda';
+import {callLambdaSync} from '../shared/call-lambda-sync';
 import {
 	getCloudwatchMethodUrl,
 	getCloudwatchRendererUrl,
@@ -88,7 +88,7 @@ export const internalRenderMediaOnLambdaRaw = async (
 	const {functionName, region, rendererFunctionName} = input;
 
 	try {
-		const res = await callLambda({
+		const res = await callLambdaSync({
 			functionName,
 			type: ServerlessRoutines.start,
 			payload: await makeLambdaRenderMediaPayload(input),
