+++ packages/lambda/src/shared/get-function-version.ts
@@ -1,7 +1,7 @@
 import type {LogLevel} from '@remotion/renderer';
 import {ServerlessRoutines} from '@remotion/serverless/client';
 import type {AwsRegion} from '../regions';
-import {callLambda} from './call-lambda';
+import {callLambdaSync} from './call-lambda-sync';
 import {COMMAND_NOT_FOUND} from './constants';
 
 export const getFunctionVersion = async ({
@@ -14,7 +14,7 @@ export const getFunctionVersion = async ({
 	logLevel: LogLevel;
 }): Promise<string> => {
 	try {
-		const result = await callLambda({
+		const result = await callLambdaSync({
 			functionName,
 			payload: {
 				logLevel,
